Reducing to only core tests - providing better test names.
