+++ metaflow/plugins/kfp/tests/sample_flows/nested_foreach.py
@@ -1,4 +1,6 @@
-from metaflow import FlowSpec, step
+# -*- coding: utf-8 -*-
+from metaflow import FlowSpec, step, Parameter, JSONType
+import pprint
 
 
 def truncate(var):
@@ -20,42 +22,11 @@ def assert_equals(expected, got):
         raise ExpectationFailed(expected, got)
 
 
-class NestedForeachTestFlow2(FlowSpec):
-    """
-    This flow contains a nested foreach and a static branch
-    on separate branches. The nested foreach and static branch
-    are joined together in `foreach_join_w_x`.
-    """
-
+class NestedForeach(FlowSpec):
     @step
     def start(self):
-        self.next(self.foreach_split_x, self.split_w)
-
-    @step
-    def foreach_split_x(self):
         self.x = "ab"
         self.next(self.foreach_split_y, foreach="x")
-
-    @step
-    def split_w(self):
-        self.var1 = 100
-        self.next(self.w1, self.w2)
-
-    @step
-    def w1(self):
-        self.var1 = 150
-        self.next(self.join_w)
-
-    @step
-    def w2(self):
-        self.var1 = 250
-        self.next(self.join_w)
-
-    @step
-    def join_w(self, w_inp):
-        self.var1 = w_inp.w1.var1
-        assert self.var1 == 150
-        self.next(self.foreach_join_w_x)
 
     @step
     def foreach_split_y(self):
@@ -69,6 +40,7 @@ class NestedForeachTestFlow2(FlowSpec):
 
     @step
     def foreach_inner(self):
+        pprint.pprint(self.input)
         [x, y, z] = self.foreach_stack()
 
         # assert that lengths are correct
@@ -82,10 +54,16 @@ class NestedForeachTestFlow2(FlowSpec):
         assert_equals(z[2], self.z[z[0]])
 
         self.combo = x[2] + y[2] + z[2]
+        self.next(self.foreach_inner_2)
+
+    @step
+    def foreach_inner_2(self):
+        assert self.input in "ef"
         self.next(self.foreach_join_z)
 
     @step
     def foreach_join_z(self, inputs):
+        pprint.pprint([(input.x, input.y, input.z) for input in inputs])
         self.next(self.foreach_join_y)
 
     @step
@@ -94,11 +72,6 @@ class NestedForeachTestFlow2(FlowSpec):
 
     @step
     def foreach_join_start(self, inputs):
-        self.next(self.foreach_join_w_x)
-
-    @step
-    def foreach_join_w_x(self, input):
-        assert input.join_w.var1 == 150
         self.next(self.end)
 
     @step
@@ -107,4 +80,4 @@ class NestedForeachTestFlow2(FlowSpec):
 
 
 if __name__ == "__main__":
-    NestedForeachTestFlow2()
+    NestedForeach()
